Changes to maps and menu
Significant overhaul to both maps (added floor coloring)
Updated button placement and text of menu, increased display quality of menu
Added font for title
updated jerks frequency to use float and changed name of slider

+++ Assets/Scripts/JerkPreferences.cs
@@ -5,5 +5,5 @@ public class JerkPreferences
 {
 	public int jerkCount = 0;
 	public float jerkWeight = 0;
-	public int jerkInterval = 0;
+	public float jerkInterval = 0;
 }
